Move well-known/webfinger to request spec
